Fix massive bug in stop-extraction algorithm
Previously, the algorithm was calculating stops across the night as one
stop-event. For example, if a taxi stopped at 10 pm on Monday, and
then started at 6 am on Tuesday, that would be an 8 hour stop. This
would have been fine, but the problem is that, after the filtering step,
some dates do not appear in the data. So a taxi might stop at 10 pm on
Monday, and the next record could be only at 6 am on Wednesday, if Tuesday was
filtered out. As a result, the stop-event would be 32 hours long! This
would be filtered out, due to the fact that I filter out unrealistically
long stops. Thus the total stop-duration on that Monday would be quite a
bit less than it's supposed to be. I've now "isolated" each day when
calculating the stop-events. The first stop event on each day will start
at 12 AM and the last stop event will end at 12 AM of the following
morning.