Re-design Hull modeling approach
The approach used be Hull's model was incorrect. The reasons for this
are to be highlighted in our paper: "Improved Electric Vehicle modelling
of minibus taxis in Africa: Addressing discrepancies in literature".

The model now uses an approach which is very similar to the one used in
SUMO (Kurczveil et al.).